requirements.txt
# Web框架
Flask==3.0.0
flask-cors==4.0.0

# 交易所API
ccxt==4.2.5

# Telegram Bot
python-telegram-bot==20.7

# 数据处理
pandas==2.1.4
numpy==1.26.3

# 数据库
SQLAlchemy==2.0.25

# 可视化
matplotlib==3.8.2
seaborn==0.13.1

# 异步处理
aiohttp==3.9.1

# 工具
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.0

# 加密
cryptography==41.0.7

# 日志
loguru==0.7.2

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
